Serve nearest_available_carpark via REST

+++ priv/repo/migrations/20220127022051_create_carparks.exs
@@ -5,8 +5,8 @@ defmodule Shopcash.Repo.Migrations.CreateCarparks do
     create table(:carparks) do
       add :number, :string, null: false
       add :address, :string, null: false
-      add :latitude, :decimal, null: false
-      add :longitude, :decimal, null: false
+      add :latitude, :float, null: false
+      add :longitude, :float, null: false
 
       timestamps()
     end
